Fix a subtle memory leak
It appears that libuv now frees the write_req->bufs struct upon a successful write.

This means that the call to `delete[] write_req->bufs[0].base;` will no longer work
because `write_req->bufs` is NULL.  The new check for `if (write_req->bufs)` added
in a previous pull request fixed the immediate segfault, but introduced
the memory leak.

To fix the issue, we now also pass along the pointer to `buf->data` in the
`write_req->data` field.